(feat share-article) Refactor code
