+++ .env-sample
@@ -4,4 +4,4 @@ REACT_APP_FACEBOOK_KEY=""
 REACT_APP_FACEBOOK_SECRET=""
 REACT_APP_TWITTER_KEY=""
 REACT_APP_TWITTER_SECRET=""
-REACT_APP_BASE_URL=
+REACT_APP_BASE_URL=herokuBackendApiUrl.com/api
